+++ Modules/Core/app/Contracts/HasInteractableComponent.php
@@ -2,34 +2,78 @@
 
 namespace Modules\Core\Contracts;
 
+use Gate;
 use Illuminate\Support\Facades\Auth;
+use Livewire\Attributes\Validate;
 
 trait HasInteractableComponent
 {
-    public function toggleLikeAction()
+    public $has_liked = false;
+
+    public function initializeHasLiked(): void
     {
-        $userId = Auth::id();
-        $ip = request()->ip();
-
-        $existingLike = $this->content->likes()
-            ->where(function ($q) use ($userId, $ip) {
-                $q->when($userId, fn ($x) => $x->where('user_id', $userId))
-                    ->when(! $userId, fn ($x) => $x->where('ip_address', $ip));
-            })
-            ->first();
-
-        if ($existingLike) {
-            $existingLike->delete();
-
-            return false;
+        if (! $this->content) {
+            return;
         }
 
-        $this->content->likes()->create([
-            'user_id' => $userId,
-            'ip_address' => $ip,
-        ]);
+        $userId = Auth::id();
+        if (! $userId) {
+            $this->has_liked = false;
 
-        return true;
+            return;
+        }
+
+        $this->has_liked = $this->content->likes()
+            ->where('user_id', $userId)
+            ->exists();
+    }
+
+    public function toggleLike()
+    {
+        if (auth()->check()) {
+            $userId = Auth::id();
+            $ip = request()->ip();
+
+            $existingLike = $this->content->likes()
+                ->where(function ($q) use ($userId, $ip) {
+                    $q->when($userId, fn ($x) => $x->where('user_id', $userId))
+                        ->when(! $userId, fn ($x) => $x->where('ip_address', $ip));
+                })
+                ->first();
+
+            if ($existingLike) {
+                $existingLike->delete();
+                $this->has_liked = false;
+
+                $this->dispatch('toastMagic',
+                    status: 'success',
+                    title: 'موفقیت',
+                    message: 'لایک حذف شد'
+                );
+                $this->refreshStats();
+
+                return;
+            }
+
+            $this->content->likes()->create([
+                'user_id' => $userId,
+                'ip_address' => $ip,
+            ]);
+            $this->has_liked = true;
+            $this->refreshStats();
+            $this->dispatch('toastMagic',
+                status: 'success',
+                title: 'موفقیت',
+                message: 'لایک شد'
+            );
+        }else{
+            $this->dispatch('toastMagic',
+                status: 'error',
+                title: 'خطا',
+                message: 'برای این عملیات باید ابتدا ثبت نام کنید'
+            );
+        }
+
     }
 
     public function visitAction()
@@ -55,12 +99,23 @@ trait HasInteractableComponent
         }
     }
 
-    public function makeCommentAction(string $text)
+    #[Validate('required|')]
+    public $commentBody;
+
+    public function makeComment()
     {
-        return $this->content->comments()->create([
+        $this->validateOnly('commentBody', [
+            'commentBody' => 'required|string|min:1|max:1000',
+        ]);
+        $this->content->comments()->create([
             'user_id' => Auth::id(),
-            'text' => $text,
+            'body' => $this->commentBody,
             'ip_address' => request()->ip(),
         ]);
+        $this->dispatch('toastMagic',
+            status: 'success',
+            title: 'موفقیت',
+            message: 'نظر با موفقیت ایجاد شد. منتظر تائید باشید'
+        );
     }
 }
